Performance fixes for multiple windows per process
The requests for the tiles for each data source are aggregated by
the DataProvider. This allows making a single asynchrounous call
to DataSource::getTileImage() in the case when multiple WallWindows
share a Tile on the same process. This guarantee removes the need
for guarding against concurrent requests in the DataProviders,
which removes lock contention and makes a better use of the thread
pool.