Bugfix: play/pause buttons did not work on movies loaded from session
The bug was introduced in #21 by using a binary copy of MovieContent via
Content::clone(). The binary copy did not (and should not) recreate the
actions, as they are already serialized by Content when sending to the
wall processes. They should only be created when making a new object on
the master process, which includes loading from xml.

Other minor cleanups, including creating the actions for
PixelStreamContents only when loading from xml to be consistent, even
though in this case the the order of the operations was different so
the problem was not visible.