Fix crash when opening movies with negative start time
The previous "invalid" value of -1 for non-video packets was in
fact still too large for a movie with a negative start time (most
extreme value encountered in a valid movie: -823638).

This caused a non-video packet to be passed to the video frame
converter, where sws_getCachedContext would abort the program on
ffmpeg < 4.0 because the packet AVPixelFormat is unkown (-1).