Fix a possible segfault with side-by-side movies
It was possible that the left & right PicturePtr& were reset
concurrently by _triggerFrameUpdate() while a getTileImage() was still
in progress, causing a segfault in left->getData(texture).

Another crash could occur when the 3D movie was looping back and not
all processes got the image, causing an invalid access in
_splitSideBySide(*image).

The first crash could be reproduced within seconds by opening a large
3D movie and moving the window around. The second was harder to
reproduce and its causes are not fully understood.

Due to the complexity of the movie decoding code there may still be a
unresolved underlying problem here; but at least this change
suppresses the crashes.