Fix InfiniBand networking by using a separate MPI channel for swap sync
The MPI standard mandates that separate threads use separate MPI
communicators. This was not the case as the render thread was doing a
swap barrier on the wall channel while the main thread was using it
for synchronizing movies and pixelstream.

This change makes Tide work over IB using IntelMPI; however mvapich2.3
still deadlocks when the two threads concurrently access the IB network.