Fix bug that could permanently block the background tap
This was due to an animation bug in Qt. Switching rapidly between the
focused and desktop modes would result in the default opacity value of
the focuscontext to be changed for a random value, making it always
visible and thus blocking the background tap on the master process.
A workaround was already in place for the wall processes but became
needed on the master as well since the latest refactorings.