Improve video playback performance (#301)
* Require ffmpeg 3.4 or later
* Improve video playback performance

The ffmpeg AVFrames are being drawn directly instead of making a copy every frame

A viewport parameter is added to the image since ffmpeg frames can have line padding. This viewport is also toggled when drawing side-by-side 3D movies.

If possible ffmpeg will decode each frame multithreaded. The number of threads to be used is set with '--decode-threads' when launching tide.